#45-feat: creating socket structure on service
